[EMCAL-790] Fix correction of the cell time for BC mod 4
The correction method used to merged the bunches
of the 4 trigger BC into 1 assumes the permutation
(0 1 2 3), otherwise it will only merge either 2 or 3 trigger
BCs into 1. Two possible issues can spoil the correction
a) Initial permutation differs from (0 1 2 3), i.e. BC 3 comes
   first, BC 0 second ...
b) Number of BCs in L0-LM correction not a multiple of 4
In order to allow for minimal intervention we treat the two
cases separately
a) We introduce a phase parameter (default 0) rotating
   the initial permutation
b) We add again L0-LM delay % 4 BCs to the BC used in
   the correction to obtain the default parameterization
In the default case we will get the permutation (0 1 2 3)
and stay with a phase of 0, non-0 phase parameters
are needed only of the BC obtained would be shifted with
respect to expectation.