Collect and send ITS/MFT raw data decoding errors
o2-itsmft-stf-decoder-workflow will send 2 extra messages (example for ITS, same for MFT):

1) {\"ITS\", \"LinkErrors\", 0} of the type std::vector<o2::itsmft::GBTLinkDecodingStat>
with GBT link decoding statistics. The GBTLinkDecodingStat is cumulated through the run but will be added to
the output of given TF only if GBTLinkDecodingStat::DecErrors type error was detected for the given link in the given TF

2) {\"ITS\", \"ChipErrors\", 0} of the type std::vector<o2::itsmft::ChipError> with Alpide decoding errors seen
in the given TF. ChipError for given chip (or feeID if the chip is not resolved, i.e. = -1) is cumulated over single
TF, providing the bit pattern of ChipStat::DecErrors and how many times this chip was in error in this TF.

To access the data in the QC:
auto linkErrors = pc.inputs().get<gsl::span<o2::itsmft::GBTLinkDecodingStat>>(\"linkerrors\");
auto detErrors = pc.inputs().get<gsl::span<o2::itsmft::ChipError>>(\"decerrors\");
for (const auto& le : linkErrors) {
  le.print();
}
for (const auto& de : decErrors) {
  LOGP(info, \"err in FeeID:{:#x} Chip:{} ErrorBits {:#b}, {} errors in TF\", de.getFEEID(), de.getChipID(), de.errors, de.nerrors);
}