[EMCAL-830] Recalibrator workflow for EMCAL cells
Applying cell level calibrations
- Bad channel masking
- Cell time
- Absolute energy
Calibrations can be enabled and disabled one-by-one
(default: all enabled). Input and output specs mut be
specifiaed by the user. Calibration objects are handeld
via the CalibLoader class, which is a helper class for
calibration object from EMCAL which can be used in DPL
tasks. In the default case the CalibLoader relies fully
on the CCDB, however for each calibration object an option
exists to load the calibration object statically from a
file (mainly test mode).