TRD: AODProducer: Fix trdPattern
In noticed sometime that 7 tracklets were counted as contributing
to the track, which is physically impossible since we only have 6
layers.

I used this function to convert from the pattern to nTracklets:

  // Count the number of contributor tracklets from the trd pattern.
  inline uint8_t nTrackletsFromPattern(uint8_t pattern) {
    return std::bitset<8>{pattern}.count();
    }

Looking at AODProducerWorkflowDPL:getTRDPattern(), we are looping
over the layers _7_ times instead of 6 (kNLayers=6 and break condition
is >=0).
In this function we call track.getTrackletIndex(6) on our first
iteration, which indexes into mAttachedTracklets[kNLayers].
This is UB!

This patch fixes it by decrementing the loop iterations by one.